Fix: Time until clocked in now works again
It would go negative until UI.getTime was called,
it wasn't being called due to only checking if we are
not inBreak, which already defaults to true.
The extra check for coChosen avoids exceptions that
would be thrown bu getTime when clock in/out time was
null, which i think fixed sheduledAtFixedRate from
randomly stopping, maybe.
- Also removed recheckTime as it was redundant.