New: WindowParent interface
Used to set windows' parent to be able to:
center on them, hide/show them, and more in the future.
Hence, no need to keep track of windows in Main, making the application
more modular and avoiding having unneeded objects in memory.
(like the break start/end windows if never needed)